Consolidate case diff logic
There is now one where there used to be two subtly different case diff implementations.
