+++ corehq/apps/couch_sql_migration/tests/test_casediff.py
@@ -1,7 +1,7 @@
 import logging
 import os
 from collections import defaultdict
-from contextlib import contextmanager
+from contextlib import ExitStack, contextmanager
 from copy import deepcopy
 from glob import glob
 from inspect import signature
@@ -33,7 +33,7 @@ class TestCaseDiffQueue(SimpleTestCase):
     def setUp(self):
         super(TestCaseDiffQueue, self).setUp()
         self.patches = [
-            patch.object(mod, "diff_cases", self.diff_cases),
+            patch.object(mod, "diff_cases_and_save_state", self.diff_cases),
             patch(
                 "corehq.form_processor.backends.couch.dbaccessors.CaseAccessorCouch.get_cases",
                 self.get_cases,
@@ -593,46 +593,45 @@ class TestDiffCases(SimpleTestCase):
 
         for patcher in self.patches:
             patcher.start()
+            self.addCleanup(patcher.stop)
         self.statedb = StateDB.init(":memory:")
+        self.addCleanup(self.statedb.close)
         self.sql_cases = {}
         self.sql_ledgers = {}
         self.couch_cases = {}
         self.couch_ledgers = {}
-
-    def tearDown(self):
-        for patcher in self.patches:
-            patcher.stop()
-        self.statedb.close()
-        super(TestDiffCases, self).tearDown()
+        stack = ExitStack()
+        stack.enter_context(mod.global_diff_state({}))
+        self.addCleanup(stack.close)
 
     def test_clean(self):
         self.add_case("a")
-        mod.diff_cases(self.couch_cases, self.statedb)
+        mod.diff_cases_and_save_state(self.couch_cases, self.statedb)
         self.assert_diffs()
 
     def test_diff(self):
         couch_json = self.add_case("a", prop=1)
         couch_json["prop"] = 2
-        mod.diff_cases(self.couch_cases, self.statedb)
+        mod.diff_cases_and_save_state(self.couch_cases, self.statedb)
         self.assert_diffs([Diff("a", path=["prop"], old=2, new=1)])
 
     def test_replace_diff(self):
         self.add_case("a", prop=1)
         different_cases = deepcopy(self.couch_cases)
         different_cases["a"]["prop"] = 2
-        mod.diff_cases(different_cases, self.statedb)
+        mod.diff_cases_and_save_state(different_cases, self.statedb)
         self.assert_diffs([Diff("a", path=["prop"], old=2, new=1)])
-        mod.diff_cases(self.couch_cases, self.statedb)
+        mod.diff_cases_and_save_state(self.couch_cases, self.statedb)
         self.assert_diffs()
 
     def test_replace_ledger_diff(self):
         self.add_case("a")
         stock = self.add_ledger("a", x=1)
         stock.values["x"] = 2
-        mod.diff_cases(self.couch_cases, self.statedb)
+        mod.diff_cases_and_save_state(self.couch_cases, self.statedb)
         self.assert_diffs([Diff("a/stock/a", "stock state", path=["x"], old=2, new=1)])
         stock.values["x"] = 1
-        mod.diff_cases(self.couch_cases, self.statedb)
+        mod.diff_cases_and_save_state(self.couch_cases, self.statedb)
         self.assert_diffs()
 
     def assert_diffs(self, expected=None):
@@ -682,8 +681,8 @@ class TestDiffCases(SimpleTestCase):
         ledgers = self.couch_ledgers
         return [ledgers[c] for c in case_id__in if c in ledgers]
 
-    def hard_rebuild_case(self, domain, case_id, *args, **kw):
-        return Config(to_json=lambda: self.couch_cases[case_id])
+    def hard_rebuild_case(self, *args, **kw):
+        raise Exception("unexpected")
 
 
 @attr.s
